pyproject.toml
[project]
name = "AetherOnePySocial"
version = "0.1.0"
description = "AetherOne Social Platform"
requires-python = ">=3.8"
dependencies = [
    "fastapi",
    "uvicorn",
    "sqlalchemy",
    "psycopg2-binary",
    "pydantic",
    "pydantic-settings",
    "email-validator",
    "passlib[bcrypt]",
    "alembic",
    "pytest",
    "sqlalchemy-to-pydantic",
    "python-jose[cryptography]",
    "python-multipart",
    "python-dotenv",
    # Test dependencies included in main dependencies
    "pytest",
    "httpx",
    "pytest-cov",
    "pytest-order",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "httpx",
    "pytest-cov",
    "pytest-order",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"] 